Switch to Postgres for database
The version of MySQL that shipped with Ubuntu had trouble
automatically creating the SPRING_SESSION_ATTRIBUTES table,
which eventually boiled down to the Spring Session code
trying to create a Varchar column that was too long when using
the UTF8MB4 encoding.

Since that has been a long-standing irritation and this is a
convenient time, trying Postgres instead for more sane UTF-8
support.

Had to modify some model classes and reports because of various
compatibility problems like reserved words and having different
functions