When taking cash, only record the amount due
The user should be giving change. IE, someone takes $80 on a $65
order - they should have given $15 back and the payment is only recorded
as $65.

For payments less than the amount due, only record the amount given.
IE, someone takes $30 on a $65 order, they are getting $0 in change
and will need to be paying more.