More at-con attendee validation
Make most required fields required (phone and email are not marked
as required because ONE of them is required. Though setting the required
parameter to disabled on the one that's not filled in might also work.

Copy emergency contact name to parent fields when Parent is Emergency
Contact box is checked