Switch to Jetty from Tomcat
Changing the application server to Jetty from Tomcat
improved performance drastically when uploading images.
With a test script that uploaded a 592 KB file 500 times,
Tomcat would average ~250ms with the maximum latency of
17,000ms+ (!)

With the same file, Jetty averaged ~185ms, with a maximum
of 221 ms.

There's quite possibly a configuration option that could be
set in Tomcat to improve performance, but I'm fine with
switching to Jetty because I use it more often at work.