Added order hand off screen/buttons
Order Hand Off allows a user to flag an at-con order for hand off, and another user to
easily find that order and resume from the same screen. The point of this is to easily move
customers to another station if there is some problem (for example, a printer dies).

Currently you can hand off from the main order screen and the print badges screen.