Rework error handling from Repository classes
For methods that are looking for a specific ID, throw a NotFoundException that will bubble up
and be handled by ErrorController.

Note that some other repository methods can legitimately return no results - for example,
AttendeeSearchRepository.searchFor() should just return an empty list if no matches are
found.

But AttendeeRepository.findById() should throw an exception, because it was searching for
a _specific_ record that wasn't there for some reason. (This should be pretty rare, but
we want to catch the edge cases and bubble up an error to be shown on the error page)